moved collect_contianer_metadata to lib, return empty dict if not compatible
